requirements.txt
requests>=2.28.0
PyYAML>=6.0
Pillow>=9.0.0
python-dateutil>=2.8.0